change app folder organization
