+++ app/src/main/java/com/zhouhaoo/sample/mvp/presenter/SettingPresenter.kt
@@ -14,27 +14,18 @@
  *  limitations under the License.
  */
 
-package com.zhouhaoo.sample.injection.module
+package com.zhouhaoo.sample.mvp.presenter
 
-import com.zhouhaoo.common.injection.ActivityScope
 import com.zhouhaoo.common.injection.FragmentScope
-import com.zhouhaoo.sample.mvp.ui.activity.MainActivity
-import com.zhouhaoo.sample.mvp.ui.fragment.SettingFragment
-import dagger.Module
-import dagger.android.ContributesAndroidInjector
+import com.zhouhaoo.common.mvp.BasePresenter
+import com.zhouhaoo.sample.mvp.contract.SettingContract
+import javax.inject.Inject
 
 /**
- * Created by zhou on 18/2/26.
+ * Created by zhou on 18/2/28.
  */
-@Module
-abstract class MainMoudle {
-    @ActivityScope
-    @ContributesAndroidInjector(modules = [MainPresenterModule::class])
-    abstract fun contributeMainActivity(): MainActivity
-
-    @FragmentScope
-    @ContributesAndroidInjector(modules = [SettingPresenterModule::class])
-    abstract fun contributeSettingFragment(): SettingFragment
-
+@FragmentScope
+class SettingPresenter @Inject constructor(model: SettingContract.Model, view: SettingContract.View) :
+        BasePresenter<SettingContract.Model, SettingContract.View>(model, view) {
 
 }
