+++ app/src/main/res/layout/activity_main.xml
@@ -21,7 +21,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.zhouhaoo.sample.features.main.MainActivity">
+    tools:context="com.zhouhaoo.sample.mvp.ui.activity.MainActivity">
 
     <Button
         android:id="@+id/tvHello"
