+++ kotlin-common/src/main/java/com/zhouhaoo/common/base/BaseFragment.kt
@@ -24,13 +24,14 @@ import android.view.ViewGroup
 import com.trello.rxlifecycle2.android.FragmentEvent
 import com.zhouhaoo.common.base.delegate.IFragment
 import com.zhouhaoo.common.integration.lifecycle.FragmentLifecycleable
+import com.zhouhaoo.common.mvp.IView
 import io.reactivex.subjects.BehaviorSubject
 import io.reactivex.subjects.Subject
 
 /**
  * Created by zhou on 18/2/27.
  */
-abstract class BaseFragment : Fragment(), IFragment, FragmentLifecycleable {
+abstract class BaseFragment : Fragment(), IFragment,IView, FragmentLifecycleable {
 
     private val mLifecycleSubject = BehaviorSubject.create<FragmentEvent>()
